Create README.md
leetcode 602 solution:
utilize the union function to count who has the most friends and the most friends number. 

(requester_id, accepter_id) is the primary key (combination of columns with unique values) for this table.
This table contains the ID of the user who sent the request, the ID of the user who received the request, and the date when the request was accepted.
 
Write a solution to find the people who have the most friends and the most friends number.

The test cases are generated so that only one person has the most friends.